Delete SmallBarStackGraph implemented in VX and implement new component to render this graph using NIVO
